pyproject.toml
[tool.poetry]
name = "ecbl"
version = "0.1.0"
description = ""
authors = ["Andrei V. Konstantinov"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">= 3.10"
python-benedict = {extras = ["yaml"], version = "^0.33.1"}
cvxpy = "^1.4.2"
scikit-learn = "^1.4.0"
einops = "^0.7.0"
torch = "^2.2.1"
pandas = "^2.1.4"
numpy = "^1.26.2"
matplotlib = "^3.8.2"
sympy = "1.12"
torchvision = "0.17.1"
torcheval = "0.0.7"
pycddlib = "2.1.7"
tqdm = "4.66.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
